Some more missing supposedly 64bit operations.
Avoid use 32bit math for extent_size.

+++ lib/activate/dev_manager.c
@@ -1412,7 +1412,7 @@ static char *_add_error_device(struct dev_manager *dm, struct dm_tree *dtree,
 	struct dm_tree_node *node;
 	struct lv_segment *seg_i;
 	int segno = -1, i = 0;
-	uint64_t size = seg->len * seg->lv->vg->extent_size;
+	uint64_t size = (uint64_t) seg->len * seg->lv->vg->extent_size;
 
 	dm_list_iterate_items(seg_i, &seg->lv->segments) {
 		if (seg == seg_i)
